+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 升级
 
 * 3.11
+  * 添加 options.upload.base64ToLink
   * 添加 options.upload.xhr
   * 添加 options.upload.cancel
 * 3.10
@@ -28,6 +29,7 @@
 
 ### v3.11.2 / 2025-06
 
+* [改进列表内粘贴列表](https://github.com/Vanessa219/vditor/pull/1830) `引入特性`
 * [改进列表内粘贴列表](https://github.com/Vanessa219/vditor/issues/1832) `改进功能`
 
 ### v3.11.1 / 2025-05-27